requirements.txt
numpy
pandas
matplotlib
plotly
python-dateutil
streamlit